Move isupport messages to handshake
