+++ test/test_helper.exs
@@ -62,6 +62,7 @@ Mimic.copy(ElixIRCd.Services.Nickserv.Verify)
 Mimic.copy(ElixIRCd.Server.Connection)
 Mimic.copy(ElixIRCd.Server.Handshake)
 Mimic.copy(ElixIRCd.Server.Dispatcher)
+Mimic.copy(ElixIRCd.Utils.Isupport)
 Mimic.copy(ElixIRCd.Utils.Mailer)
 Mimic.copy(ElixIRCd.Utils.Network)
 Mimic.copy(ElixIRCd.Utils.Nickserv)
